requirements.txt
Django
psycopg2
djangorestframework
django-rest-auth
django-allauth
drf-extensions
markdown
pygments
coreapi
drf-haystack
geopy
elasticsearch==2.4.1
factory_boy